Move OBBBA structural reforms to benefits to current policy Fixes #6313
