+++ policyengine_us/parameters/gov/aca/ineligible_immigration_statuses.yaml
@@ -5,6 +5,15 @@ values:
     - UNDOCUMENTED
     - TPS 
 
+  2027-01-01:
+    - DACA_TPS #daca is eligible for PTC
+    - UNDOCUMENTED
+    - TPS 
+    - ASYLEE
+    - DEPORTATION_WITHHELD
+    - PAROLED_ONE_YEAR
+
+
 metadata:
   label: Reconciliation Premium Tax Credit immigration status adjustment in effect
   period: year
@@ -12,3 +21,5 @@ metadata:
   reference:
     - title: 26 CFR § 1.36B-1 - Premium tax credit definitions.
       href: https://www.law.cornell.edu/cfr/text/26/1.36B-1#(g)
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
