+++ policyengine_us/parameters/gov/hhs/medicaid/eligibility/work_requirements/applies.yaml
@@ -1,0 +1,13 @@
+description: The the Department of Health and Human Services limits Medicaid to filers who meet work requirements, if this is true. 
+
+values:
+  0000-01-01: false
+  2027-01-01: true
+
+metadata:
+  unit: bool
+  period: year
+  label: Medicaid work requirements apply
+  reference:
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
