+++ policyengine_us/parameters/gov/hhs/medicaid/eligibility/work_requirements/dependent_age_limit.yaml
@@ -1,0 +1,12 @@
+description: The the Department of Health and Human Services exempts Medicaid work requirement to individuals with a dependent child at this age or younger.
+
+values:
+  2027-01-01: 13
+
+metadata:
+  unit: year
+  label: Medicaid work requirement dependent child age limit
+  period: year
+  reference:
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
