+++ policyengine_us/parameters/gov/hhs/medicaid/eligibility/work_requirements/monthly_hours_threshold.yaml
@@ -1,0 +1,12 @@
+description: The the Department of Health and Human Services limits Medicaid to individuals working more than this number of monthly hours.
+
+values:
+  2027-01-01: 80
+
+metadata:
+  unit: hours
+  label: Medicaid work requirement monthly hours threshold
+  period: month
+  reference:
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
