+++ policyengine_us/parameters/gov/usda/snap/work_requirements/abawd/age_threshold/dependent.yaml
@@ -1,6 +1,7 @@
 description: The United States Department of Agriculture considers individuals who do not have any dependents under this age to be Able-Bodied Adult Without Dependents (ABAWD) under the Supplemental Nutrition Assistance Program.
 values:
   2019-01-01: 18
+  2027-01-01: 14
   
 metadata:
   unit: year
@@ -11,4 +12,5 @@ metadata:
       href: https://www.law.cornell.edu/cfr/text/7/273.1#b_iii
     - title: SNAP Work Requirements
       href: https://www.fns.usda.gov/snap/work-requirements
-  
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
