+++ policyengine_us/parameters/gov/usda/snap/work_requirements/abawd/age_threshold/exempted.yaml
@@ -10,6 +10,8 @@ metadata:
       href: https://www.law.cornell.edu/cfr/text/7/273.24#c_1
     - title: SNAP Work Requirements
       href: https://www.fns.usda.gov/snap/work-requirements
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
 brackets:
   - threshold:
       2019-01-01: 0
@@ -21,5 +23,6 @@ brackets:
       2019-01-01: false
   - threshold: 
       2019-01-01: 55
+      2027-01-01: 65
     amount: 
       2019-01-01: true
