+++ policyengine_us/parameters/gov/usda/snap/work_requirements/abawd/amended_structure/in_effect.yaml
@@ -1,0 +1,13 @@
+description: The United States Department of Agriculture applies the amended structure to the Able-Bodied Adult Without Dependents (ABAWD) work requirements for the Supplemental Nutrition Assistance Program, if this is true.
+
+values:
+  0000-01-01: false
+  2027-01-01: true
+
+metadata:
+  unit: bool
+  period: year
+  label: SNAP ABAWD work requirements amended structure in effect
+  reference:
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
