+++ policyengine_us/parameters/gov/usda/snap/work_requirements/abawd/amended_structure/sunset_provisions_apply.yaml
@@ -1,0 +1,13 @@
+description: The United States Department of Agriculture applies the sunset provisions to the Able-Bodied Adult Without Dependents (ABAWD) work requirements for the Supplemental Nutrition Assistance Program, if this is true.
+
+values:
+  0000-01-01: false
+  2030-10-01: true
+
+metadata:
+  unit: bool
+  period: year
+  label: SNAP ABAWD work requirements sunset provisions apply
+  reference:
+    - title: H.R.1 - One Big Beautiful Bill Act
+      href: https://www.congress.gov/bill/119th-congress/house-bill/1/text
