+++ policyengine_us/tests/policy/baseline/gov/hhs/medicaid/eligibility/medicaid_work_requirement_eligible.yaml
@@ -1,0 +1,88 @@
+- name: Case 1, age 18, work exempted age. 
+  period: 2026
+  input:
+    age: 18
+  output:
+    medicaid_work_requirement_eligible: true
+
+- name: Case 2, age 65, work exempted age. 
+  period: 2026
+  input:
+    age: 65
+  output:
+    medicaid_work_requirement_eligible: true
+
+- name: Case 3, age 19, need working. 
+  period: 2026
+  input:
+    age: 19
+  output:
+    medicaid_work_requirement_eligible: false
+
+- name: Case 4, age 30, working 79 hours per month. 
+  period: 2026
+  input:
+    age: 30
+    monthly_hours_worked: 79
+  output:
+    medicaid_work_requirement_eligible: false
+
+- name: Case 5, age 30, working 80 hours per month. 
+  period: 2026
+  input:
+    age: 30
+    monthly_hours_worked: 80
+  output:
+    medicaid_work_requirement_eligible: true
+
+- name: Case 6, age 30, is disabled. 
+  period: 2026
+  input:
+    age: 19
+    monthly_hours_worked: 5
+    is_disabled: true
+  output:
+    medicaid_work_requirement_eligible: true
+
+- name: Case 7, integration. 
+  period: 2026
+  input:
+    people:
+      person1:
+        age: 30
+        is_tax_unit_head_or_spouse: true
+        monthly_hours_worked: 80
+      person2:
+        age: 30
+        is_tax_unit_head_or_spouse: true
+        monthly_hours_worked: 79
+  output:
+    medicaid_work_requirement_eligible: [true, false]
+
+- name: Case 8, not working, but has child at age 13. 
+  period: 2026
+  input:
+    people:
+      person1:
+        age: 30
+        is_tax_unit_head_or_spouse: true
+        monthly_hours_worked: 0
+      person2:
+        age: 13
+        is_tax_unit_dependent: true
+  output:
+    medicaid_work_requirement_eligible: [true, true]
+
+- name: Case 9, not working, has child at age 14, ineligible under Senate Version. 
+  period: 2026
+  input:
+    people:
+      person1:
+        age: 30
+        is_tax_unit_head_or_spouse: true
+        monthly_hours_worked: 0
+      person2:
+        age: 14
+        is_tax_unit_dependent: true
+  output:
+    medicaid_work_requirement_eligible: [false, true]
