--- policyengine_us/tests/policy/contrib/reconciliation/reconciled_snap_abawd_work_requirement.yaml
@@ -1,118 +1,0 @@
-- name: Case 1, baseline, age 55 exempted from general work requirement. 
-  period: 2026
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: false
-    age: 55
-    weekly_hours_worked: 1
-    is_disabled: false
-  output:
-    meets_snap_abawd_work_requirements: true
-
-- name: Case 2, reform, age 55 is not exempted from general work requirement. 
-  period: 2027
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: true
-    age: 55
-    weekly_hours_worked: 1
-    is_disabled: false
-  output:
-    meets_snap_abawd_work_requirements: false
-
-- name: Case 3, baseline, have a child at age 15.
-  period: 2026-01
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: false
-    people:
-      person1:
-        age: 55
-        weekly_hours_worked: 1
-        is_disabled: false
-        is_parent: true
-      person2:
-        age: 15
-        is_disabled: false
-        is_tax_unit_dependent: true
-  output:
-    meets_snap_abawd_work_requirements: [true, true]
-
-- name: Case 4, reform, have a child at age 15.
-  period: 2027
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: true
-    people:
-      person1:
-        age: 55
-        weekly_hours_worked: 1
-        is_disabled: false
-        is_parent: true
-      person2:
-        age: 15
-        is_disabled: false
-        is_tax_unit_dependent: true
-  output:
-    meets_snap_abawd_work_requirements: [false, true]
-
-- name: Case 5, reform, have a child at age 6.
-  period: 2027
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: true
-    people:
-      person1:
-        age: 55
-        weekly_hours_worked: 1
-        is_disabled: false
-        is_parent: true
-      person2:
-        age: 6
-        is_disabled: false
-        is_tax_unit_dependent: true
-  output:
-    meets_snap_abawd_work_requirements: [true, true]
-
-- name: Case 6, reform, is veteran with a child 15 in 2027. 
-  period: 2027
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: true
-    people:
-      person1:
-        age: 55
-        weekly_hours_worked: 1
-        is_disabled: false
-        is_veteran: true
-      person2:
-        age: 15
-        is_disabled: false
-        is_tax_unit_dependent: true
-  output:
-    meets_snap_abawd_work_requirements: [true, true]
-
-- name: Case 7, reform, is veteran with a child 15 in 2031, sunset provision in effect. 
-  period: 2031
-  absolute_error_margin: 0.9
-  reforms: policyengine_us.reforms.reconciliation.reconciled_snap_abawd_work_requirement.reconciled_snap_abawd_work_requirement
-  input:
-    gov.contrib.reconciliation.snap_abawd_work_requirement.in_effect: true
-    people:
-      person1:
-        age: 55
-        weekly_hours_worked: 1
-        is_disabled: false
-        is_veteran: true
-      person2:
-        age: 15
-        is_disabled: false
-        is_tax_unit_dependent: true
-  output:
-    meets_snap_abawd_work_requirements: [false, true]
