+++ policyengine_us/variables/gov/usda/snap/eligibility/work_requirements/meets_snap_abawd_work_requirements.py
@@ -14,16 +14,16 @@ class meets_snap_abawd_work_requirements(Variable):
         weekly_hours_worked = person("weekly_hours_worked", period.this_year)
         # Work at least 20 hours a week
         is_working = weekly_hours_worked >= p.weekly_hours_threshold
-        # Under 18 or 55 years of age or older are exempted
+        # Adults within an age range are exempt
         worked_exempted_age = p.age_threshold.exempted.calc(age)
         # Unable to work due to a physical or mental limitation
         is_disabled = person("is_disabled", period)
         # Parent of a household member under 18
         is_dependent = person("is_tax_unit_dependent", period)
-        is_child = age < p.age_threshold.dependent
+        is_qualifying_child = age < p.age_threshold.dependent
         is_parent = person("is_parent", period)
-        has_child = person.spm_unit.any(is_dependent & is_child)
-        exempted_parent = is_parent & has_child
+        has_child = person.spm_unit.any(is_dependent & is_qualifying_child)
+        exempt_parent = is_parent & has_child
         # Exempted from the general work requirements
         meets_snap_general_work_requirements = person(
             "meets_snap_general_work_requirements", period
@@ -34,11 +34,41 @@ class meets_snap_abawd_work_requirements(Variable):
         is_homeless = person.household("is_homeless", period)
         # A veteran
         is_veteran = person("is_veteran", period)
+        if p.amended_structure.in_effect:
+            is_child = person("is_child", period)
+            older_child_dependent = (
+                is_dependent & is_child & ~is_qualifying_child
+            )
+            is_married = person.family("is_married", period)
+            has_older_child_dependent = person.spm_unit.any(
+                older_child_dependent
+            )
+            is_head_or_spouse = person("is_tax_unit_head_or_spouse", period)
+            has_head_or_spouse_working = person.spm_unit.any(
+                is_head_or_spouse & is_working
+            )
+            exempt_married_person = where(
+                is_married & has_older_child_dependent,
+                has_head_or_spouse_working,
+                0,
+            )
+            base_conditions = (
+                is_working
+                | worked_exempted_age
+                | is_disabled
+                | exempt_parent
+                | meets_snap_general_work_requirements
+                | is_pregnant
+                | exempt_married_person
+            )
+            if p.amended_structure.sunset_provisions_apply:
+                return base_conditions
+            return base_conditions | is_homeless | is_veteran
         return (
             is_working
             | worked_exempted_age
             | is_disabled
-            | exempted_parent
+            | exempt_parent
             | meets_snap_general_work_requirements
             | is_pregnant
             | is_homeless
